Call Python constructors for intent(OUT) arguments. Call PyCapsule_New for void pointers.

+++ src/components/shroud/tests/example/wrapExClass2.cpp
@@ -136,7 +136,7 @@ int AA_exclass2_get_type_id(AA_exclass2 * self)
 ExClass2 *selfobj = static_cast<ExClass2 *>(static_cast<void *>(self));
 // splicer begin class.ExClass2.method.get_type_id
 TypeID rv = selfobj->getTypeID();
-return rv;
+return static_cast<int>(rv);
 // splicer end class.ExClass2.method.get_type_id
 }
